uc: provide Windows build
This add/changes the bits needed for a Windows build, either by build
tags (when Linux and Darwin) are in the same file, or using the correct
file suffix (i.e. _windows.go for windows stuff). This is done exactly
as done for the few Darwin stubs (also made Darwin uppercase 'D'
everywhere).

Thus this slightly re-arranges the code and adds a few stubs for
Windows, without making any source code level changes.

See #209

Note this *just* provides a Window and does not implement a socks proxy
for Windows (if that is even necessary).

~~~
% GOOS=windows go build && ls uc.exe
uc.exe
% file uc.exe
uc.exe: PE32+ executable for MS Windows 6.01 (console), ARM64, 14 sections
% GOOS=windows GOARCH=amd64 go build && ls uc.exe
% file uc.exe
uc.exe: PE32+ executable for MS Windows 6.01 (console), x86-64, 16 sections
~~~
