+++ internal/machine/docker/controller_windows.go
@@ -6,14 +6,14 @@ import (
 	"net/netip"
 )
 
-// EnsureUncloudNetwork is a stub for Darwin.
+// EnsureUncloudNetwork is a stub for Windows.
 func (c *Controller) EnsureUncloudNetwork(
 	ctx context.Context, subnet netip.Prefix, mtu int, dnsServer netip.Addr,
 ) error {
-	return fmt.Errorf("not supported on Darwin")
+	return fmt.Errorf("not supported on Windows")
 }
 
-// Cleanup is a stub for Darwin.
+// Cleanup is a stub for Windows.
 func (c *Controller) Cleanup() error {
-	return fmt.Errorf("not supported on Darwin")
+	return fmt.Errorf("not supported on Windows")
 }
