+++ internal/machine/network/mtu_linux.go
@@ -1,5 +1,3 @@
-//go:build linux
-
 package network
 
 import (
