+++ internal/machine/network/mtu_windows.go
@@ -1,0 +1,8 @@
+package network
+
+import "errors"
+
+// detectEgressMTU is a stub for Windows. The machine daemon that performs detection only runs on Linux.
+func detectEgressMTU() (int, error) {
+	return 0, errors.New("not implemented on Windows")
+}
