+++ internal/machine/unix_socket_windows.go
@@ -1,0 +1,8 @@
+package machine
+
+import "net"
+
+// listenUnixSocket is not supported on Windows.
+func listenUnixSocket(path string) (net.Listener, error) {
+	return nil, nil
+}
